cloud_run_proj/pyproject.toml
[project]
name = "cloud-run-proj"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
    "yfinance>=0.2.65",
]
